# apps/web/docs/openapi.yaml
openapi: 3.0.3
info:
  title: StudioOS API
  description: Enterprise Interior Design Platform API
  version: 1.0.0
  contact:
    name: StudioOS Support
    email: support@studioos.com
    url: https://studioos.com/support
  license:
    name: Proprietary
    url: https://studioos.com/license
servers:
  - url: https://api.studioos.com/v1
    description: Production server
  - url: https://staging-api.studioos.com/v1
    description: Staging server
  - url: https://dev-api.studioos.com/v1
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /projects:
    get:
      summary: List projects
      description: Get a paginated list of projects for the authenticated user
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/SortBy'
        - $ref: '#/components/parameters/SortOrder'
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, active, completed, archived]
        - name: client_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Projects retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create project
      description: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Project created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '400':
          description: Invalid project data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page
    SortBy:
      name: sort_by
      in: query
      schema:
        type: string
        enum: [created_at, updated_at, name, status]
        default: created_at
      description: Field to sort by
    SortOrder:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: desc
      description: Sort order
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
        timestamp:
          type: string
          format: date-time
    ProjectsResponse:
      type: object
      properties:
        projects:
          type: array
          items:
            $ref: '#/components/schemas/Project'
        pagination:
          $ref: '#/components/schemas/Pagination'
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [draft, active, completed, archived]
        client_id:
          type: string
          format: uuid
        budget:
          type: number
        deadline:
          type: string
          format: date
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProjectResponse:
      type: object
      properties:
        project:
          $ref: '#/components/schemas/Project'
    CreateProjectRequest:
      type: object
      required:
        - name
        - client_id
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 1000
        client_id:
          type: string
          format: uuid
        budget:
          type: number
          minimum: 0
        deadline:
          type: string
          format: date
        tags:
          type: array
          items:
            type: string
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        total_pages:
          type: integer
          minimum: 1